+++ packages/std/src/math/mod.rs
@@ -1,9 +1,11 @@
 mod decimal;
 mod fraction;
+mod isqrt;
 mod uint128;
 mod uint64;
 
 pub use decimal::Decimal;
 pub use fraction::Fraction;
+pub use isqrt::Isqrt;
 pub use uint128::Uint128;
 pub use uint64::Uint64;